app/requirements.txt
# fastapi
# uvicorn
# pymongo
# bcrypt==4.0.1
# passlib[bcrypt]==1.7.4
# python-jose
# python-dotenv
# email-validator



fastapi
uvicorn
pymongo
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose
python-dotenv
email-validator
fastapi-mail